Correcting Windows build brokenness.

+++ source/core_io_reader.cpp
@@ -185,7 +185,7 @@ CoreIOReader::ReadLine ( char *_buffer, int _bufferLength )
 	m_ioMutex->Unlock ();
 #endif
 
-	return bytesRead;
+	return (int)bytesRead;
 }
 
 int
